update.sh: Add warning for development
When users make local changes to submodule(s) for testing/development, it's up to them to fix merge conflicts.
Don't let this script undo their local changes.
It could be misleading.
Actually the 'merge conflicts Yes/No question' is superfluous with OV 9.2 as we don't patching submodules, and OV 9.2
is still under development.