Introduce DEVELOPER_NAME for installing specific stuff in our test images.
I have my own plugin called Persian Palace + its own android app which will be ready soon.

It's not on private/public git (it's a local plugin) so for better recipe management I just introduced this variable so we could add what we want as test images are not official builds on a build server.