Remove extra WLANs from feeds like FOREVER!
Such a fucked up way to detect what could be built and what couldn't which is wrong almost all the time so get over this shit.

How many issues/topics/reports should we get about this and when this madness will be over?

All internal drivers added to the metas so anything else is extra and needed to be tested against all kernels, something that I will never have the time for it to do.

We have another variable KERNELVERSION which is static in each meta and could help if somebody wants to come up with a real solution for this otherwise this will be always disabled.