Add ipv6 kernel module to all images and autoload it by default
Seems some brands/models need this module for network connections.

It's a module in all Open Vision models except wetekplay as without internal it won't boot.

We should consider this so there may be more models like wetekplay we don't know about.

But it's better to have this as module to not let anything go default by IPv6 and don't create big kernels.